:truck: rename: 수정된 my 위젯 경로로 import 수정
